requirements.txt
# 采集层
python-jobspy>=1.1.0        # LinkedIn/Indeed 聚合采集
httpx>=0.27.0               # The Hub API 请求

# 智能层 (Google Vertex AI)
google-cloud-aiplatform>=1.40.0   # Embeddings (Legacy/Standard SDK)
google-genai>=0.3.0               # Gemini 3 Pro + Thinking (New V2 SDK)
numpy>=1.26.0                     # 余弦相似度计算

# 生成层
jinja2>=3.1.0               # LaTeX 模板渲染
pyyaml>=6.0                 # profile.yaml 解析
python-dotenv>=1.0.0        # 读取 .env 环境变量

# 动态网页抓取
playwright>=1.40              # Company career page rendering (JS)

# 测试
pytest>=7.0                   # Test framework
pytest-asyncio>=0.21          # Async test support

# Tectonic 需要系统安装，不是 pip 包
# macOS: brew install tectonic
# Linux: cargo install tectonic
